refactor: rename `token_to_user.py` to `middlewares.py`
