+++ quickstart/quickstart/schema.py
@@ -1,45 +1,27 @@
 # quickstart.schema.py
 
 import strawberry
-from strawberry_django_plus import gql
 from strawberry_django_plus.directives import SchemaDirectiveExtension
-from strawberry_django_plus.permissions import IsAuthenticated
 
-from gqlauth.core.directives import IsVerified
 from gqlauth.user import arg_mutations as mutations
 from gqlauth.user.queries import UserQueries
 from gqlauth.user.resolvers import Captcha
 
 
 @strawberry.type
-class MyAuthorizedQueries(UserQueries):
-    @strawberry.field
-    def secured_string(self) -> str:
-        return "secure"
+class Query(UserQueries):
+    ...
 
 
 @strawberry.type
-class Query:
-    @gql.django.field(directives=[IsVerified()])
-    def auth_entry(self) -> MyAuthorizedQueries:
-        return MyAuthorizedQueries()
-
-
-@strawberry.type
-class AuthMutation:
+class Mutation:
+    # these require to be validated
     verify_token = mutations.VerifyToken.field
     update_account = mutations.UpdateAccount.field
     archive_account = mutations.ArchiveAccount.field
     delete_account = mutations.DeleteAccount.field
     password_change = mutations.PasswordChange.field
-
-
-@strawberry.type
-class Mutation:
-    @gql.django.field(directives=[IsAuthenticated()])
-    def auth_entry(self) -> AuthMutation:
-        return AuthMutation()
-
+    # these can be used by anonymous users.
     captcha = Captcha.field
     token_auth = mutations.ObtainJSONWebToken.field
     register = mutations.Register.field
