+++ testproject/asgi.py
@@ -3,7 +3,7 @@ from channels.routing import ProtocolTypeRouter, URLRouter
 from django.urls import re_path
 from strawberry.channels import GraphQLHTTPConsumer, GraphQLWSConsumer
 
-from gqlauth.core.token_to_user import ChannelsJwtMiddleware
+from gqlauth.core.middlewares import ChannelsJwtMiddleware
 from testproject.schema import arg_schema
 
 websocket_urlpatterns = [
